requirements.txt
fastapi
uvicorn
mysql-connector-python
sqlmodel
SQLAlchemy-Utils
python-dotenv
pandas
xlrd